Correct inverted condition in test
When trying to drain the queue it was looping on retrieving messages
until it got something.  So if the queue was already empty and not being
filled it would wait forever, and if there were multiple messages in the
queue it wouldn't actually be drained.

Switch to getting messages until no more are found.  Won't guarantee the
queue remains empty if other processes are writing to it, but it will at
least work with a private server and will likely often work with the
public one used as the default.